1st commit - uploading of extra files
Heres the extra files used for the project, as of now this mainly consists of two files. A jupyter notebook containing the code that helps me go through a json file and fetch the appropriate data from the file. This is then exported to a simple csv file that can then be shared between the group. This will also include the scryfall file, seeing as it will be usefull for anyone who wishes to recreate that part of the project. 
The data files are included in a zip due to the large file size of both of these.